test: break controller test into two suites, reduce sleep times
